# [Refactor] Use the new optimize function