--- a/src/benchmark/nam_small_circuits.cpp
+++ b/src/benchmark/nam_small_circuits.cpp
@@ -3,6 +3,7 @@
 #include "quartz/tasograph/substitution.h"
 #include <cmath>
 #include <filesystem>
+#include <sstream>
 
 using namespace quartz;
 
@@ -12,6 +13,7 @@ static const std::string
 
 static int num_benchmark = 0;
 static double geomean_gate_count = 1;
+static std::stringstream summary_result;
 
 void benchmark_nam(const std::string &circuit_name) {
   std::string circuit_path = "circuit/nam-benchmarks/" + circuit_name + ".qasm";
@@ -55,19 +57,12 @@ void benchmark_nam(const std::string &circuit_name) {
 
   start = std::chrono::steady_clock::now();
   // Optimization
-  auto graph_after_search = graph_before_search->optimize(
-      1.0001,
-      0,
-      false,
-      &dst_ctx,
-      ecc_set_name, /*use_simulated_annealing=*/
-      false, /*enable_early_stop=*/
-      false,
-      /*use_rotation_merging_in_searching=*/
-      false,
-      GateType::rz,
-      circuit_name, /*timeout=*/
-      10);
+  auto graph_after_search = graph_before_search->optimize(&dst_ctx,
+                                                          ecc_set_name,
+                                                          circuit_name, /*print_message=*/
+                                                          true, /*cost_upper_bound=*/
+                                                          -1, /*timeout=*/
+                                                          10);
   end = std::chrono::steady_clock::now();
 
   std::cout << circuit_name << " optimization result in "
@@ -78,6 +73,15 @@ void benchmark_nam(const std::string &circuit_name) {
             << " seconds: gate count = " << graph_after_search->gate_count()
             << ", circuit depth = " << graph_after_search->circuit_depth()
             << ", cost = " << graph_after_search->total_cost() << std::endl;
+
+  summary_result << circuit_name << " optimization result in "
+                 << (double) std::chrono::duration_cast<std::chrono::milliseconds>(
+                     end - start)
+                     .count() /
+                     1000.0
+                 << " seconds: gate count = " << graph_after_search->gate_count()
+                 << ", circuit depth = " << graph_after_search->circuit_depth()
+                 << ", cost = " << graph_after_search->total_cost() << std::endl;
 
   geomean_gate_count /= graph_after_search->gate_count();
   // graph_after_search->to_qasm(output_fn, false, false);
@@ -97,6 +101,8 @@ int main() {
   benchmark_nam("vbe_adder_3");  // 150 gates
   benchmark_nam("gf2^4_mult");  // 225 gates
   if (num_benchmark > 0) {
+    std::cout << "Summary:" << std::endl;
+    std::cout << summary_result.str();
     std::cout << num_benchmark << " circuits, gate count optimized by "
               << std::pow(geomean_gate_count, 1.0 / num_benchmark)
               << " times on average."