--- a/src/quartz/tasograph/tasograph.h
+++ b/src/quartz/tasograph/tasograph.h
@@ -201,15 +201,28 @@ public:
                                        RuleParser *rule_parser,
                                        bool ignore_toffoli = false);
   std::shared_ptr<Graph>
-  optimize(float alpha, int budget, bool print_subst, Context *ctx,
-           const std::string &equiv_file_name, bool use_simulated_annealing,
-           bool enable_early_stop, bool use_rotation_merging_in_searching,
-           GateType target_rotation, std::string circuit_name = "",
-           int timeout = 86400 /*1 day*/);
-  std::shared_ptr<Graph> optimize(std::vector<GraphXfer *> xfers,
-                                  double gate_count_upper_bound,
-                                  std::string circuit_name, bool print_message,
-                                  int timeout = 86400 /*1 day*/);
+  optimize_legacy(float alpha,
+                  int budget,
+                  bool print_subst,
+                  Context *ctx,
+                  const std::string &equiv_file_name,
+                  bool use_simulated_annealing,
+                  bool enable_early_stop,
+                  bool use_rotation_merging_in_searching,
+                  GateType target_rotation,
+                  std::string circuit_name = "",
+                  int timeout = 86400 /*1 day*/);
+  std::shared_ptr<Graph> optimize(Context *ctx,
+                                  const std::string &equiv_file_name,
+                                  const std::string &circuit_name,
+                                  bool print_message,
+                                  double cost_upper_bound = -1 /*default = current cost * 1.05*/,
+                                  int timeout = 3600 /*1 hour*/);
+  std::shared_ptr<Graph> optimize(const std::vector<GraphXfer *> &xfers,
+                                  double cost_upper_bound,
+                                  const std::string &circuit_name,
+                                  bool print_message,
+                                  int timeout = 3600 /*1 hour*/);
   void constant_and_rotation_elimination();
   void rotation_merging(GateType target_rotation);
   std::string to_qasm(bool print_result, bool print_id) const;