--- a/src/test/test_ibmq.cpp
+++ b/src/test/test_ibmq.cpp
@@ -81,13 +81,10 @@ int main(int argc, char **argv) {
   }
 
   // Optimization
-  auto graph_after_search = graph_before_search->optimize(
-      0.999, 0, false, &dst_ctx, eqset_fn, simulated_annealing, early_stop,
-      /*rotation_merging_in_searching*/ false, GateType::u1);
-  //  eqset_fn = "../IBM_with_U3_2_1_complete_ECC_set.json";
-  //  Graph *graph_after_u3_pass = graph_after_search->optimize(
-  //      0.999, 0, false, &dst_ctx, eqset_fn, simulated_annealing, early_stop,
-  //      /*rotation_merging_in_searching*/ false, GateType::u1);
+  auto graph_after_search = graph_before_search->optimize(&dst_ctx,
+                                                          eqset_fn,
+                                                          fn, /*print_message=*/
+                                                          true);
   end = std::chrono::steady_clock::now();
   std::cout << "Optimization results of Quartz for " << fn
             << " on IBMQ gate set." << std::endl