--- a/src/test/test_rigetti_td_disabled.cpp
+++ b/src/test/test_rigetti_td_disabled.cpp
@@ -25,6 +25,7 @@ int main(int argc, char **argv) {
   bool early_stop = false;
   parse_args(argv, argc, simulated_annealing, early_stop, input_fn, output_fn,
              eqset_fn);
+  auto fn = input_fn.substr(input_fn.rfind('/') + 1);
 
   // Construct contexts
   Context src_ctx({GateType::h, GateType::ccz, GateType::cx, GateType::x,
@@ -58,10 +59,10 @@ int main(int argc, char **argv) {
   auto union_ctx_0 = union_contexts(&cz_ctx, &dst_ctx);
   auto graph_before_h_cz_merge = new_graph->context_shift(
       &dst_ctx, &cz_ctx, &union_ctx_0, &cx_2_cz, false);
-  auto graph_after_h_cz_merge = graph_before_h_cz_merge->optimize(
-      0.999, 0, false, &union_ctx_0, "../H_CZ_2_2_complete_ECC_set.json",
-      simulated_annealing, false, /*rotation_merging_in_searching*/ true,
-      GateType::rz);
+  auto graph_after_h_cz_merge = graph_before_h_cz_merge->optimize(&union_ctx_0,
+                                                                  "../H_CZ_2_2_complete_ECC_set_modified.json",
+                                                                  fn, /*print_message=*/
+                                                                  true);
   graph_after_h_cz_merge->to_qasm(
       "circuit/voqc-benchmarks/after_h_cz_merge.qasm", false, false);
 
@@ -75,19 +76,19 @@ int main(int argc, char **argv) {
       &cz_ctx, &rigetti_ctx, &union_ctx_1, &rules, false);
 
   // Optimization
-  auto graph_after_search = graph_rigetti->optimize(
-      0.999, 0, false, &union_ctx_1, eqset_fn, simulated_annealing, early_stop,
-      /*rotation_merging_in_searching*/ false, GateType::rz);
+  auto graph_after_search = graph_rigetti->optimize(&union_ctx_1,
+                                                    eqset_fn,
+                                                    fn, /*print_message=*/
+                                                    true);
   auto end = std::chrono::steady_clock::now();
-  auto fn = input_fn.substr(input_fn.rfind('/') + 1);
   std::cout << "Optimization results of Quartz for " << fn
             << " on Rigetti gate set." << std::endl
             << "Gate count after optimization: "
             << graph_after_search->total_cost() << ", "
-            << (double)std::chrono::duration_cast<std::chrono::milliseconds>(
-                   end - start)
-                       .count() /
-                   1000.0
+            << (double) std::chrono::duration_cast<std::chrono::milliseconds>(
+                end - start)
+                .count() /
+                1000.0
             << " seconds." << std::endl;
   graph_after_search->to_qasm(output_fn, false, false);
 }