This PR refactors the codebase to use the new `optimize` function in all cases, and rename the old one to be `optimize_legacy`.

It also creates a `optimize` function to take as input the ECC set file name and call the new `optimize` function for easier invocation.

It also adds a short summary for `src/benchmark/nam_small_circuits.cpp` for it to be easier to read.

It also changes the default timeout from 1 day to 1 hour.

Benchmark on Ubuntu workstation:
Before this PR:
```
tof_3 optimization result in 10.277 seconds: gate count = 35, cost = 35
barenco_tof_3 optimization result in 10.381 seconds: gate count = 40, cost = 40
mod_mult_55 optimization result in 10.849 seconds: gate count = 97, cost = 97
vbe_adder_3 optimization result in 10.977 seconds: gate count = 109, cost = 109
gf2^4_mult optimization result in 19.473 seconds: gate count = 182, cost = 182
5 circuits, gate count optimized by 1.08384 times on average.
```
After this PR:
```
tof_3 optimization result in 10.194 seconds: gate count = 35, circuit depth = 31, cost = 35
barenco_tof_3 optimization result in 10.497 seconds: gate count = 40, circuit depth = 37, cost = 40
mod_mult_55 optimization result in 10.717 seconds: gate count = 96, circuit depth = 46, cost = 96
vbe_adder_3 optimization result in 10.563 seconds: gate count = 92, circuit depth = 58, cost = 92
gf2^4_mult optimization result in 10.632 seconds: gate count = 184, circuit depth = 103, cost = 184
5 circuits, gate count optimized by 1.1211 times on average.
```