@xumingkuan 2022-08-03
I found that we will run out of memory when running for a long time (e.g., 1 hour). This is also an issue with the old `optimize` function, although it has been fixed in the artifact repo. I will merge this for now and create another PR to fix it like the artifact repo.